% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.ts.DEEM.R
\name{est.ts.DEEM}
\alias{est.ts.DEEM}
\title{Calculate the DEEM estimator and its estimated standard error in the two-sample setting}
\usage{
est.ts.DEEM(
  index,
  Gamma_h_lst,
  gamma_h_lst,
  gamma_s_lst,
  Sig_y,
  Sig_x,
  Sig_alpha,
  R_sel
)
}
\arguments{
\item{index}{Number of LD blocks included.}

\item{Gamma_h_lst}{A list of vectors of estimated SNP–outcome associations within different LD blocks from the outcome sample.}

\item{gamma_h_lst}{A list of vectors of estimated SNP–exposure associations within different LD blocks from the exposure sample.}

\item{gamma_s_lst}{A list of vectors of estimated SNP–exposure associations within different LD blocks from the supplemental exposure sample (used to select SNPs).}

\item{Sig_y}{A list of estimated variance-covariance matrix for the estimated SNP–outcome associations within different LD blocks from the outcome sample.}

\item{Sig_x}{A list of estimated variance-covariance matrix for the estimated SNP–exposure associations within different LD blocks from the exposure sample.}

\item{Sig_alpha}{A list of estimated variance-covariance matrix for the pleiotropic effects within different LD blocks.}

\item{R_sel}{A list of estimated LD matrix within different LD blocks.}
}
\value{
A vector containing:
\itemize{
\item \code{beta_hat}: the estimated causal effect.
\item \code{se}: the estimated standard error.
}
}
\description{
This function implements the Debiased Estimating Equation Method (DEEM) for Mendelian
randomization in the two-sample setting, producing consistent causal effect estimates even when instruments are weak
and/or correlated.
Inputs of this function are the outputs of the function "prepare.DEEM".
Examples for the required summary statistics can be downloaded from https://osf.io/j25rc/files/osfstorage
(the LD matrix is approximately 6GB after decompression).
}
