% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.DEEM.R
\name{prepare.DEEM}
\alias{prepare.DEEM}
\title{Prepare the data required by DEEM, outputs of this function are the inputs of "est.ts.DEEM" and "est.os.DEEM"}
\usage{
prepare.DEEM(
  SumData,
  com_list,
  LD_Ref,
  Blocks,
  ns,
  ne,
  no,
  p_thr = 0.1,
  r2_thr = 0.81
)
}
\arguments{
\item{SumData}{A data.frame consists of eight variables:
\itemize{
\item \code{chr}: the chromosome.
\item \code{pos}: the position.
\item \code{gamma_s}: a vector of estimated SNP–exposure associations from the supplemental exposure sample used to select the SNPs.
\item \code{sigma_s}: a vector of estimated standard error for the estimated SNP–exposure associations from the supplemental exposure sample used to select the SNPs.
\item \code{gamma_h}: a vector of estimated SNP–exposure associations from the exposure sample.
\item \code{sigma_x}: a vector of estimated standard error for the estimated SNP–exposure associations from the exposure sample.
\item \code{Gamma_h}: a vector of estimated SNP–outcome associations from the outcome sample.
\item \code{sigma_y}: a vector of estimated standard error for the estimated SNP–outcome associations from the outcome sample.
}
SNPs in the three samples should be the same (possible preprocessing required).}

\item{com_list}{A vector of characters in the form "chr:pos" of the SNPs in the SumData.}

\item{LD_Ref}{A list of LD matrix in each LD blocks from the reference sample.}

\item{ns}{sample size of the supplemental exposure sample that generates gamma_s.}

\item{ne}{sample size of the exposure sample that generates gamma_h.}

\item{no}{sample size of the outcome sample that generates Gamma_h.}

\item{p_thr}{p-value threshold, default 0.1.}

\item{r2_thr}{r2 threshold, default 0.81.}

\item{Block}{data.frame consists of three variables:
\itemize{
\item \code{chr}: a vector of character in the form "chr?" where "?" is the chromosome that the LD block locates on.
\item \code{start}: a vector of int indicates where the LD block starts.
\item \code{stop}: a vector of int indicates where the LD block ends.
}}
}
\value{
A list containing:
\itemize{
\item \code{index}: Number of LD blocks included.
\item \code{Gamma_h_lst}: A list of vectors of estimated SNP–outcome associations within different LD blocks from the outcome sample.
\item \code{gamma_h_lst}: A list of vectors of estimated SNP–exposure associations within different LD blocks from the exposure sample.
\item \code{gamma_s_lst}: A list of vectors of estimated SNP–exposure associations within different LD blocks from the supplemental exposure sample (used to select SNPs).
\item \code{Sig_y}: A list of estimated variance-covariance matrix for the estimated SNP–outcome associations within different LD blocks from the outcome sample.
\item \code{Sig_x}: A list of estimated variance-covariance matrix for the estimated SNP–exposure associations within different LD blocks from the exposure sample.
\item \code{Sig_alpha}: A list of estimated variance-covariance matrix for the pleiotropic effects within different LD blocks.
\item \code{R_sel}: A list of estimated LD matrix within different LD blocks.
}
}
\description{
The function prepares the data required by DEEM, outputs of this function are the inputs of "est.ts.DEEM"
and "est.os.DEEM"
Examples for the required summary statistics can be downloaded from https://osf.io/j25rc/files/osfstorage
(the LD matrix is approximately 6GB after decompression)
}
